finished adding the tags page
