+++ src/app/(tags)/tag/[id]/page.tsx
@@ -1,0 +1,30 @@
+import { getTagById } from "@/lib/backendFunction/getTagbyId";
+import { getQuestionsByTag } from "@/lib/backendFunction/getQuestionsByTag";
+import { QuestionWithRelations } from "@/lib/backendFunction/getAllQuestion";
+
+import PostCard from "@/app/components/PostCard";
+
+export default async function TagPage({ params }: { params: { id: string } }) {
+    const tagId = params.id;
+    const tag = await getTagById(tagId);
+    const questions = (await getQuestionsByTag(
+        tagId
+    )) as QuestionWithRelations[];
+
+    if (!tag) {
+        return <div>Tag not found</div>;
+    }
+
+    return (
+        <div>
+            <h1>{tag.name}</h1>
+            <p>{tag.description}</p>
+            <h1 className="text-2xl font-bold text-white mt-8">Posts</h1>
+            <div className="flex flex-col gap-4">
+                {questions.map((q) => (
+                    <PostCard key={q.id} question={q} />
+                ))}
+            </div>
+        </div>
+    );
+}
