+++ src/lib/backendFunction/getQuestionsByTag.ts
@@ -1,36 +1,33 @@
 import prisma from "@/lib/prisma";
-import { Prisma } from "@prisma/client";
 
-export type QuestionModelWithCount = Prisma.QuestionGetPayload<{
-    include: {
-        poster: { select: { id: true; name: true; email: true } };
-        tag: { select: { id: true; name: true; color: true } };
-    };
-    _count: {
-        answers: true;
-    };
-}>;
+import { QuestionWithRelations } from "./getAllQuestion";
 
 export async function getQuestionsByTag(
     tagId: string
-): Promise<QuestionModelWithCount[] | null> {
+): Promise<QuestionWithRelations[] | null> {
     if (!tagId) {
         console.error("Tag ID is required");
         return null;
     }
-    const questions = await prisma.question.findMany({
+    const posts = await prisma.question.findMany({
         where: {
             tagid: tagId,
         },
+        orderBy: { createdAt: "desc" },
         include: {
             poster: { select: { id: true, name: true, email: true } },
             tag: { select: { id: true, name: true, color: true } },
+            answers: true,
         },
     });
 
-    return questions.map((p) => ({
+    if (!posts) {
+        return null;
+    }
+
+    return posts.map((p) => ({
         ...p,
         title: p.title ?? "",
         description: p.description ?? "",
-    }));
+    })) as QuestionWithRelations[];
 }
